Split ast stdlib module into files
