+++ vm/src/stdlib/ast/node.rs
@@ -1,0 +1,63 @@
+use crate::{PyObjectRef, PyResult, VirtualMachine};
+
+pub(crate) trait Node: Sized {
+    fn ast_to_object(self, vm: &VirtualMachine) -> PyObjectRef;
+    fn ast_from_object(vm: &VirtualMachine, object: PyObjectRef) -> PyResult<Self>;
+}
+
+impl<T: Node> Node for Vec<T> {
+    fn ast_to_object(self, vm: &VirtualMachine) -> PyObjectRef {
+        vm.ctx
+            .new_list(
+                self.into_iter()
+                    .map(|node| node.ast_to_object(vm))
+                    .collect(),
+            )
+            .into()
+    }
+
+    fn ast_from_object(vm: &VirtualMachine, object: PyObjectRef) -> PyResult<Self> {
+        vm.extract_elements_with(&object, |obj| Node::ast_from_object(vm, obj))
+    }
+}
+
+impl<T: Node> Node for Box<T> {
+    fn ast_to_object(self, vm: &VirtualMachine) -> PyObjectRef {
+        (*self).ast_to_object(vm)
+    }
+
+    fn ast_from_object(vm: &VirtualMachine, object: PyObjectRef) -> PyResult<Self> {
+        T::ast_from_object(vm, object).map(Box::new)
+    }
+}
+
+impl<T: Node> Node for Option<T> {
+    fn ast_to_object(self, vm: &VirtualMachine) -> PyObjectRef {
+        match self {
+            Some(node) => node.ast_to_object(vm),
+            None => vm.ctx.none(),
+        }
+    }
+
+    fn ast_from_object(vm: &VirtualMachine, object: PyObjectRef) -> PyResult<Self> {
+        if vm.is_none(&object) {
+            Ok(None)
+        } else {
+            Ok(Some(T::ast_from_object(vm, object)?))
+        }
+    }
+}
+
+pub(super) struct BoxedSlice<T>(pub(super) Box<[T]>);
+
+impl<T: Node> Node for BoxedSlice<T> {
+    fn ast_to_object(self, vm: &VirtualMachine) -> PyObjectRef {
+        self.0.into_vec().ast_to_object(vm)
+    }
+
+    fn ast_from_object(vm: &VirtualMachine, object: PyObjectRef) -> PyResult<Self> {
+        Ok(Self(
+            <Vec<T> as Node>::ast_from_object(vm, object)?.into_boxed_slice(),
+        ))
+    }
+}
